make changes on makefile

+++ Makefile
@@ -1,4 +1,4 @@
-init: build up composer-install-prestissimo composer-install-app set-storage-link clear-cache dump-autoload copy-env migrate npm-prod queue-on
+init: build up composer-install-prestissimo composer-install-app set-storage-link clear-cache dump-autoload copy-env migrate npm-install npm-prod
 
 build:
 	docker-compose build
@@ -10,14 +10,16 @@ down:
 	docker-compose down
 
 composer-install-prestissimo:
-	mv src/composer.json runtime/composer.json
-	mv src/composer.lock runtime/composer.lock
+	if [ -d "./runtime" ]; then rm -rf runtime; fi
+	mkdir runtime
+	if [ -f "./src/composer.json" ]; then mv src/composer.json runtime/composer.json; fi
+	if [ -f "./src/composer.lock" ]; then mv src/composer.lock runtime/composer.lock; fi
 	cp sources/composer.json src/composer.json
 	echo -e "\e[1minstall composer booster\e[0m"
 	docker-compose run --rm --no-deps php-fpm composer require hirak/prestissimo
-	rm src/composer.lock
-	mv runtime/composer.json src/composer.json
-	mv runtime/composer.lock src/composer.lock
+	if [ -f "./src/composer.lock" ]; then rm -f src/composer.lock; fi
+	if [ -f "./runtime/composer.json" ]; then mv runtime/composer.json src/composer.json; fi
+	if [ -f "./runtime/composer.lock" ]; then mv runtime/composer.lock src/composer.lock; fi
 	echo -e "\e[1;37;42minstall composer booster...............................done\e[0m"
 
 composer-install-app:
@@ -30,8 +32,6 @@ set-storage-link:
 	docker-compose run --rm --no-deps php-fpm php artisan storage:link
 
 queue-on:
-	echo -e "\e[1;37;42mready to use\e[0m"
-	echo -e "\e[1mdocs: https://github.com/Steamvis/laravel-crm/tree/master/docs\e[0m"
 	docker-compose run --rm --no-deps php-fpm php artisan queue:work
 
 migrate:
@@ -46,9 +46,17 @@ dump-autoload:
 clear-logs:
 	docker-compose run --rm --no-deps php-fpm rm -rf storage/logs/laravel.log
 
+npm-install:
+	if [ -d "./src/node_modules" ]; then rm -rf ./src/node_modules; fi
+	if [ -f "./src/package-lock.json" ]; then rm -f ./src/package-lock.json; fi
+	docker-compose run --rm --no-deps node npm cache clean --force
+	docker-compose run --rm --no-deps node npm i
+
 npm-prod:
-	docker-compose run --rm --no-deps node npm install --save-dev webpack
+	if [ -d "./runtime" ]; then rm -rf runtime; fi
 	docker-compose run --rm --no-deps node npm run prod
+	echo -e "\e[1;37;42mready to use\e[0m"
+	echo -e "\e[1mdocs: https://github.com/Steamvis/laravel-crm/tree/master/docs\e[0m"
 
 copy-env:
 	cp sources/env-laravel src/.env
